requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
black==22.8.0
click==8.1.3
Django==4.1.1
django-cors-headers==3.13.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-spectacular==0.24.2
importlib-resources==5.9.0
inflection==0.5.1
jsonschema==4.16.0
mypy-extensions==0.4.3
pathspec==0.10.1
phonenumbers==8.12.56
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
PyJWT==2.5.0
pyrsistent==0.18.1
pytz==2022.4
PyYAML==6.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.3.0
uritemplate==4.1.1
zipp==3.8.1
